# Remove barrier in Metal matmul