@geohot 2023-05-23
I found the barrier is faster on M1 Max, slower on plain M1.