@philipturner 2023-05-23
Thanks for the insight. A lot of small things can be like that. For example, unrolling the same loop sometimes makes it faster, sometimes worse.