@geohot 2023-05-23
Yea, the real way to solve this is to parameterize the search space and let your smartest search algo find the winner.