@philipturner 2023-05-23
There's a simpler solution to the problem, where you can maximize performance with minimal tuning. Might be up the alley of Tinygrad.

https://arxiv.org/abs/2301.03598