@geohot 2023-05-25
Going to close. Will read that paper.